backend/requirements.txt
Flask==3.1.3
Flask-CORS==6.0.2
opencv-python==4.13.0.92
torch==2.11.0
torchvision==0.26.0
pillow==12.2.0
numpy==2.4.4
tensorflow==2.21.0
keras==3.14.0
gunicorn==25.3.0
python-dotenv==1.2.2
requests==2.33.1